requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
docker==7.1.0
idna==3.10
pyaes==1.6.1
pyrotgfork==2.1.35
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
TgCrypto==1.2.5
urllib3==2.5.0
